sort yolo classifier to vision to save code lines from main service
