+++ src/main/java/org/myrobotlab/inmoov/Vision.java
@@ -1,13 +1,24 @@
 package org.myrobotlab.inmoov;
 
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashMap;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Set;
+import java.util.TreeMap;
 import java.util.TreeSet;
 
+import javax.swing.JFrame;
+
+import org.myrobotlab.document.Classification;
 import org.myrobotlab.logging.LoggerFactory;
 import org.myrobotlab.opencv.OpenCVFilter;
 import org.myrobotlab.service.InMoov;
 import org.myrobotlab.service.OpenCV;
+import org.myrobotlab.service.Runtime;
+import org.myrobotlab.service.SwingGui;
 import org.slf4j.Logger;
 
 /**
@@ -19,6 +30,10 @@ public class Vision {
   transient public InMoov instance;
   public final static Logger log = LoggerFactory.getLogger(Vision.class);
   public Set<String> preFilters = new TreeSet<String>();
+  transient public LinkedHashMap<String, Double> collectionPositions = new LinkedHashMap<String, Double>();
+  transient public HashMap<String, Integer> collectionCount = new HashMap<String, Integer>();
+  transient private HashMap<String, Integer> collectionTemp = new HashMap<String, Integer>();
+  transient public Set<String> filteredLabel = new TreeSet<String>();
 
   /**
    * Init default config parameters
@@ -58,6 +73,9 @@ public class Vision {
     preFilters.forEach(name -> instance.opencv.addFilter(name).enable());
   }
 
+  /**
+   * setActiveFilter will add filter if no exist, enable and setactive
+   */
   public OpenCVFilter setActiveFilter(String filterName) {
     if (!Arrays.asList(OpenCV.getPossibleFilters()).contains(filterName)) {
       log.error("Sorry, {} is an unknown filter.", filterName);
@@ -68,6 +86,11 @@ public class Vision {
       instance.opencv.addFilter(filterName);
     }
     instance.opencv.setActiveFilter(filterName);
+    SwingGui gui = (SwingGui) Runtime.getService("gui");
+    //fix overexpand windows
+    if (gui != null) {
+      gui.getFrame().setExtendedState(JFrame.MAXIMIZED_BOTH);
+    }
     return (OpenCVFilter) instance.opencv.getFilter(filterName);
   }
 
@@ -76,7 +99,6 @@ public class Vision {
       instance.opencv.capture();
       if (instance.opencv.isCapturing()) {
         enablePreFilters();
-        instance.opencv.stopCapture();
         ready = true;
         return true;
       }
@@ -109,5 +131,48 @@ public class Vision {
    */
   public boolean isReady() {
     return ready;
+  }
+
+  /**
+   * sub onClassification to a analyze yolo filter, and save some code lines from main service
+   */
+  public void onClassification(TreeMap<String, List<Classification>> classifications) {
+    
+    //reset previous same classified objects ( to count same objetcs on the frame )
+    for (Map.Entry<String, List<Classification>> entry : classifications.entrySet()) {
+      List<Classification> value = entry.getValue();
+      for (Classification document : value) {
+        if (collectionCount.containsKey(document.getLabel())) {
+          collectionCount.remove(document.getLabel());
+        }
+      }
+    }
+
+    //add now labels and positions to collection
+    for (Map.Entry<String, List<Classification>> entry : classifications.entrySet()) {
+      List<Classification> value = entry.getValue();
+      for (Classification document : value) {
+        // sometime we want to filter labels
+        if (!filteredLabel.contains(document.getLabel())) {
+          Integer existingLabelCount = 1;
+          if (collectionCount.containsKey(document.getLabel())) {
+            existingLabelCount = collectionCount.get(document.getLabel()) + 1;
+          }
+          collectionCount.put(document.getLabel(), existingLabelCount);
+          collectionPositions.put(document.getLabel(), (double) document.getBoundingBox().x);
+        }
+      }
+    }
+    //Sort result based on position
+    collectionTemp.clear();
+    collectionPositions.entrySet().stream().sorted(Map.Entry.comparingByValue()).forEach(key -> collectionTemp.put(key.getKey(), collectionCount.get(key.getKey())));
+    collectionCount.clear();
+    collectionCount.putAll(collectionTemp);
+    log.info("onClassification collectionCount : {}", collectionTemp);
+  }
+
+  public Integer getPosition(String text) {
+    List<String> indexes = new ArrayList<String>(collectionCount.keySet());
+    return indexes.indexOf(text) + 1;
   }
 }
